Move marquee view to separate file
